Split out a couple mutations
