+++ src/server/implementation/types.ts
@@ -127,92 +127,3 @@ export type Doc<T extends TableNamesInDataModel<AuthDataModel>> = GenericDoc<
   AuthDataModel,
   T
 >;
-
-export const storeArgs = {
-  args: v.union(
-    v.object({
-      type: v.literal("signIn"),
-      userId: v.id("users"),
-      sessionId: v.optional(v.id("authSessions")),
-      generateTokens: v.boolean(),
-    }),
-    v.object({
-      type: v.literal("signOut"),
-    }),
-    v.object({
-      type: v.literal("refreshSession"),
-      refreshToken: v.string(),
-    }),
-    v.object({
-      type: v.literal("verifyCodeAndSignIn"),
-      params: v.any(),
-      provider: v.optional(v.string()),
-      verifier: v.optional(v.string()),
-      generateTokens: v.boolean(),
-      allowExtraProviders: v.boolean(),
-    }),
-    v.object({
-      type: v.literal("verifier"),
-    }),
-    v.object({
-      type: v.literal("verifierSignature"),
-      verifier: v.string(),
-      signature: v.string(),
-    }),
-    v.object({
-      type: v.literal("userOAuth"),
-      provider: v.string(),
-      providerAccountId: v.string(),
-      profile: v.any(),
-      signature: v.string(),
-    }),
-    v.object({
-      type: v.literal("createVerificationCode"),
-      accountId: v.optional(v.id("authAccounts")),
-      provider: v.string(),
-      email: v.optional(v.string()),
-      phone: v.optional(v.string()),
-      code: v.string(),
-      expirationTime: v.number(),
-      allowExtraProviders: v.boolean(),
-    }),
-    v.object({
-      type: v.literal("createAccountFromCredentials"),
-      provider: v.string(),
-      account: v.object({ id: v.string(), secret: v.optional(v.string()) }),
-      profile: v.any(),
-      shouldLinkViaEmail: v.optional(v.boolean()),
-      shouldLinkViaPhone: v.optional(v.boolean()),
-    }),
-    v.object({
-      type: v.literal("retrieveAccountWithCredentials"),
-      provider: v.string(),
-      account: v.object({ id: v.string(), secret: v.optional(v.string()) }),
-    }),
-    v.object({
-      type: v.literal("modifyAccount"),
-      provider: v.string(),
-      account: v.object({ id: v.string(), secret: v.string() }),
-    }),
-    v.object({
-      type: v.literal("invalidateSessions"),
-      userId: v.id("users"),
-      except: v.optional(v.array(v.id("authSessions"))),
-    }),
-  ),
-};
-
-export const internal: {
-  auth: {
-    store: FunctionReference<
-      "mutation",
-      "internal",
-      ObjectType<typeof storeArgs>,
-      any
-    >;
-  };
-} = {
-  auth: {
-    store: "auth:store" as any,
-  },
-};
